refactor(ui): delete_confirm et suppression des méthodes "Controller::remove()"

+++ controller/ControllerComment.php
@@ -11,9 +11,12 @@ class ControllerComment extends ExtendedController {
 
     public function delete() {
         $comment = $this->get_object_or_redirect("id", "Comment");
-        $this->authorize_for_board_or_redirect($comment->get_board());
+        $user = $this->authorize_for_board_or_redirect($comment->get_board());
 
-        CommentDao::delete($comment);
+        if ($user->can_delete_comment($comment)) {
+            CommentDao::delete($comment);
+        }
+
         $this->redirect("card", "view", $comment->get_card_id());
     }
 
@@ -24,7 +27,6 @@ class ControllerComment extends ExtendedController {
             $board = $comment->get_board();
             $user = $this->authorize_for_board_or_redirect($board);
 
-
             (new View("comment_edit"))->show(array(
                     "user" => $user,
                     "comment" => $comment,
@@ -32,7 +34,8 @@ class ControllerComment extends ExtendedController {
                     "errors" => Session::get_error()
                 )
             );
-        } else {
+        }
+        else {
             $comment = $this->get_object_or_redirect("id", "Comment");
             $this->authorize_for_board_or_redirect($comment->get_board());
 
@@ -47,38 +50,16 @@ class ControllerComment extends ExtendedController {
         }
     }
 
-    public function edit_confirm() {
-        $comment = $this->get_object_or_redirect("id", "Comment");
-        $this->authorize_for_board_or_redirect($comment->get_board());
-
-        if(Post::all_sets("validate", "body")){
-            $body = Post::get("body");
-            $comment->set_body($body);
-            CommentDao::update($comment);
-        }
-
-       $this->card_redirect($comment->get_card_id());
-    }
-
     public function add(){
         $card = $this->get_object_or_redirect("card_id", "Card");
         $user = $this->authorize_for_board_or_redirect($card->get_board());
 
-
         // si 'body' est vide, ne fait rien, pas besoin de message
+        // TODO: quand-même vérifier que pas que espaces
         if(!Post::empty("body")) {
             $comment = new Comment(Post::get("body"), $user, $card);
             CommentDao::insert($comment);
         }
-        $this->card_redirect($card->get_id());
-    }
-
-
-    private function card_redirect($card_id) {
-        if(Post::isset("edit")){
-            $this->redirect("card", "edit", $card_id);
-        } else {
-            $this->redirect("card", "view", $card_id);
-        }
+        $this->redirect("card", "view", $comment->get_card_id());
     }
 }
